rm address encoding
