+++ test/fixtures.json
@@ -1,100 +1,4 @@
 {
-  "addresses": {
-    "valid": [
-      {
-        "string": "BC1QW508D6QEJXTDG4Y5R3ZARVARY0C5XW7KV8F3T4",
-        "prefix": "bc",
-        "hex": "751e76e8199196d454941c45d1b3a323f1433bd6",
-        "version": 0
-      },
-      {
-        "string": "tb1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3q0sl5k7",
-        "prefix": "tb",
-        "hex": "1863143c14c5166804bd19203356da136c985678cd4d27a1b8c6329604903262",
-        "version": 0
-      },
-      {
-        "string": "bc1pw508d6qejxtdg4y5r3zarvary0c5xw7kw508d6qejxtdg4y5r3zarvary0c5xw7k7grplx",
-        "prefix": "bc",
-        "hex": "751e76e8199196d454941c45d1b3a323f1433bd6751e76e8199196d454941c45d1b3a323f1433bd6",
-        "version": 1
-      },
-      {
-        "string": "BC1SW50QA3JX3S",
-        "prefix": "bc",
-        "hex": "751e",
-        "version": 16
-      },
-      {
-        "string": "bc1zw508d6qejxtdg4y5r3zarvaryvg6kdaj",
-        "prefix": "bc",
-        "hex": "751e76e8199196d454941c45d1b3a323",
-        "version": 2
-      },
-      {
-        "string": "tb1qqqqqp399et2xygdj5xreqhjjvcmzhxw4aywxecjdzew6hylgvsesrxh6hy",
-        "prefix": "tb",
-        "hex": "000000c4a5cad46221b2a187905e5266362b99d5e91c6ce24d165dab93e86433",
-        "version": 0
-      }
-    ],
-    "invalid": [
-      {
-        "ignore": true,
-        "string": "tc1qw508d6qejxtdg4y5r3zarvary0c5xw7kg3g4ty",
-        "exception": "Invalid network"
-      },
-      {
-        "string": "bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t5",
-        "exception": "Invalid checksum"
-      },
-      {
-        "string": "BC13W508D6QEJXTDG4Y5R3ZARVARY0C5XW7KN40WF2",
-        "prefix": "bc",
-        "version": 17,
-        "program": "751e76e8199196d454941c45d1b3a323f1433bd6",
-        "exception": "Invalid version \\(17\\)"
-      },
-      {
-        "string": "BC1SW50QA3JX3S",
-        "prefix": "foo",
-        "exception": "Expected foo, got bc"
-      },
-      {
-        "string": "bc1rw5uspcuh",
-        "prefix": "bc",
-        "version": 1,
-        "program": "75",
-        "exception": "Program too short"
-      },
-      {
-        "string": "bc10w508d6qejxtdg4y5r3zarvary0c5xw7kw508d6qejxtdg4y5r3zarvary0c5xw7kw5rljs90",
-        "prefix": "bc",
-        "version": 1,
-        "program": "751e76e8199196d454941c45d1b3a323f1433bd6751e76e8199196d454941c45d1b3a323f1433bd675",
-        "exception": "Program too long"
-      },
-      {
-        "string": "BC1QR508D6QEJXTDG4Y5R3ZARVARYV98GJ9P",
-        "prefix": "bc",
-        "version": 0,
-        "program": "1d1e76e8199196d454941c45d1b3a323",
-        "exception": "Unknown program"
-      },
-      {
-        "string": "tb1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3q0sL5k7",
-        "exception": "Mixed-case string"
-      },
-      {
-        "string": "tb1pw508d6qejxtdg4y5r3zarqfsj6c3",
-        "exception": "Excess padding"
-      },
-      {
-        "string": "tb1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3pjxtptv",
-        "exception": "Non-zero padding"
-      }
-    ]
-  },
   "fromWords": {
     "invalid": [
       {
